Extract value registry
